+++ starsky/starsky.foundation.sync/WatcherBackgroundService/DiskWatcherBackgroundTaskQueue.cs
@@ -1,0 +1,11 @@
+using starsky.foundation.injection;
+using starsky.foundation.worker.Services;
+
+namespace starsky.foundation.sync.WatcherBackgroundService
+{
+	[Service(typeof(DiskWatcherBackgroundTaskQueue), InjectionLifetime = InjectionLifetime.Singleton)]
+	public class DiskWatcherBackgroundTaskQueue : BackgroundTaskQueue
+	{
+		// is empty due inheritance
+	}
+}