+++ starsky/starsky.foundation.sync/WatcherServices/DiskWatcherBackgroundService.cs
@@ -35,7 +35,11 @@ namespace starsky.foundation.sync.WatcherServices
 		{
 			_logger.LogInformation((_appSettings.UseDiskWatcher != false ? $"UseDiskWatcher is enabled" 
 				: "UseDiskWatcher is disabled" ) + $" on {Environment.MachineName}");
-			if ( _appSettings.UseDiskWatcher == false ) return Task.CompletedTask;
+			if ( _appSettings.UseDiskWatcher == false
+			     && _appSettings.ApplicationType == AppSettings.StarskyAppType.WebController )
+			{
+				return Task.CompletedTask;
+			}
 
 			_diskWatcher.Watcher(_appSettings.StorageFolder);
 			return Task.CompletedTask;