+++ starsky/starsky/Controllers/SyncController.cs
@@ -15,7 +15,7 @@ using starsky.foundation.realtime.Interfaces;
 using starsky.foundation.storage.Interfaces;
 using starsky.foundation.storage.Models;
 using starsky.foundation.storage.Storage;
-using starsky.foundation.worker.Services;
+using starsky.foundation.worker.Interfaces;
 using starskycore.Interfaces;
 using starskycore.ViewModels;