+++ starsky/starskytest/starsky.foundation.worker/BackgroundTaskQueueTest.cs
@@ -13,6 +13,7 @@ using starsky.foundation.platform.Extensions;
 using starsky.foundation.platform.Interfaces;
 using starsky.foundation.platform.Models;
 using starsky.foundation.webtelemetry.Interfaces;
+using starsky.foundation.worker.Interfaces;
 using starsky.foundation.worker.Services;
 using starskytest.FakeMocks;